Multiple Networks

+++ src/wallet/api.rs
@@ -7,8 +7,7 @@ use grin_wallet::libwallet::types::{
     OutputStatus, Context, NodeClient, OutputData, TxLogEntry, TxLogEntryType, WalletBackend,
 };
 use grin_wallet::libwallet::{Error, ErrorKind};
-use grin_core::core::Transaction;
-use grin_core::core::amount_to_hr_string;
+use grin_core::core::{amount_to_hr_string, Transaction};
 use grin_keychain::{Keychain, Identifier};
 use grin_core::libtx::slate::Slate;
 use grin_core::libtx::{build, tx_fee};
@@ -499,7 +498,7 @@ fn select_send_tx<T: ?Sized, C, K>(
             if coins.len() == max_outputs {
                 return Err(ErrorKind::NotEnoughFunds {
                     available: total as u64,
-                    available_disp: amount_to_hr_string(total as u64, false),
+                    available_disp: amount_to_hr_string(total, false),
                     needed: amount_with_fee as u64,
                     needed_disp: amount_to_hr_string(amount_with_fee as u64, false),
                 })?;
